updated messages hooks
